requirements.txt
pandas==2.0.0
numpy==1.24.3
requests==2.31.0
sentence-transformers==2.2.2
scikit-learn==1.2.2
faiss-cpu==1.7.4
hdbscan==0.8.29
streamlit==1.25.0 